refactor: vacation service imports and update tsconfig.json
